Fixed some signed/unsigned comparisons
darcs-hash:20080323235614-fafb9-ad25482fe68beae207e8d89efeb0f28f2b40108c.gz

+++ itkHierarchicalQueue.h
@@ -153,7 +153,7 @@ public:
   /** push a value in the queue */
   inline void Push( const KeyType & k, const ValueType & v)
     {
-    assert( k  - NT::NonpositiveMin() < m_Vector.size() );
+      assert( (int)(k  - NT::NonpositiveMin()) < (int)m_Vector.size() );
     assert( k  - NT::NonpositiveMin() >= 0 );
 
     m_Vector[ k  - NT::NonpositiveMin() ].push_back( v );
